pyproject.toml
[tool.poetry]
name = "knowledge-graph"
version = "0.0.1"
description = ""
authors = ["dadevel@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
markdown = "^3.4.1"
pygments = "^2.13.0"
pymdown-extensions = "^9.9"
jinja2 = "^3.1.2"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
